Support templates with two indices
This is needed for things like pwm1_auto_point2_temp.

The code is updated to keep track of number of times a template is
instantiated and a macros are added to allow setting that count.
Then nct6687_create_attr_group() is taught to allocate enough memory and
run one more nested loop for the second index.